Attempted to update predictions.
Added a broken piece of after the working prediction gsp. The code
creates a radio button for every possible match, some that start
disabled if they are not in the first round.

I attempted to have the disabled radio buttons update as the matches
were filled in, however I'm running into a null error and cannot
progress until I discover a solution.